Add getGradesByStudentId, getAllStudentsWIthGrades
Enhancements made to the system involve the incorporation of a novel
endpoint, namely "/students/:id/grades," within the updated
studentsControllerV2. This strategic addition facilitates the seamless
invocation of the recently introduced function
named getGradesByStudentId.
The principal purpose of this function is to retrieve and furnish
all academic grades associated with a specific student, contingent
upon their unique student ID.

Furthermore, a pioneering function titled getAllStudentsWithGrades
has been introduced. This innovative function operates in a manner
such that when the endpoint is designated as
"/students?includes=grades," it proficiently returns an encompassing
collection of students, accompanied by a meticulously organized array
of their respective grades. This augmentation significantly enhances
the system's functionality, rendering it more efficient and versatile
in handling student data and grade-related inquiries.